backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
openai==1.107.2
python-dotenv==1.1.1
langchain-openai==0.3.33
langgraph==0.6.6
sqlalchemy==2.0.39
alembic==1.15.2
psycopg[binary]==3.2.6
PyJWT[crypto]==2.10.1
